feat: implementar agendamiento de citas con autenticación JWT y actualización de entidades
Cambios en autenticación:
- Refactorizado AuthController para usar AuthInput como objeto de entrada GraphQL
- Corregido orden de claims en JwtTokenProvider (setClaims antes de setSubject)
- Creado AuthInput DTO para encapsular credenciales de login

Implementación de agendamiento de citas:
- Agregado CitaExamenController (renombrado desde AgendamientoGraphQLController)
- Creado AgendamientoInput DTO para capturar datos del cliente
- Implementado CitaExamenService con lógica de agendamiento y validaciones
- Agregado DisponibilidadService para gestionar consultas de disponibilidad

Actualización de entidades y repositorios:
- Modificado Disponibilidad: agregados métodos ocuparCupo(), liberarCupo() y tieneCuposDisponibles()
- Actualizado TipoExamen: cambio de Integer a Long en ID, agregada relación @OneToMany con cascada
- Actualizado DisponibilidadRepository: incluido parámetro horaInicio en findAndLockDisponibilidad
- Agregado método findByNombreIgnoreCase en TipoExamenRepository
- Cambiados tipos de ID de Integer a Long en múltiples repositorios para consistencia

Ajustes de configuración:
- Actualizado puerto de conexión a base de datos (5432 → 6543) en application.properties
- Limpiados imports no utilizados en SecurityConfig

Estas modificaciones permiten el agendamiento completo de citas de exámenes médicos
con validación de disponibilidad, control de cupos y autenticación basada en JWT.

+++ src/main/java/com/CitaSalud/domain/repository/SedeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SedeRepository extends JpaRepository<Sede, Integer> {
+public interface SedeRepository extends JpaRepository<Sede, Long> {
 }
